document-transformer/requirements.txt
# Markdown和HTML处理
markdown==3.5.2
beautifulsoup4==4.12.2
jinja2==3.1.2
html2text==2020.1.16

# PowerPoint处理
python-pptx==0.6.21

# Word文档处理
python-docx==0.8.11
nltk==3.8.1
pandas==2.0.3
openpyxl==3.1.2

# 配置和工具
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

PyPDF2==3.0.1
langchain==0.0.350
matplotlib==3.7.1
tqdm==4.65.0
numpy==1.26.4
scipy==1.15.2
markdown-it-py==2.2.0
markdown-it-charts==0.1.0
pillow==9.5.0 